@@ -35,6 +35,7 @@ import android.graphics.RectF;
 import android.graphics.drawable.LayerDrawable;
 import android.os.Build;
 import android.os.Bundle;
+import android.os.Process;
 import android.support.annotation.NonNull;
 import android.support.v4.app.LoaderManager;
 import android.support.v4.content.CursorLoader;
@@ -420,53 +421,50 @@ public class CalendarFragment2 extends CalendarPermissionFragment implements
 
         public void loadPeriod(final int periodIndex, final boolean forceIt) {
 
-            new Runnable() {
-                @Override
-                public void run() {
-                    android.os.Process.setThreadPriority(android.os.Process.THREAD_PRIORITY_BACKGROUND);
+            ((Runnable) () -> {
+                Process.setThreadPriority(Process.THREAD_PRIORITY_BACKGROUND);
 
-                    int index = mLastLoadedPeriods.indexOf(periodIndex);
-                    if (index < 0 || forceIt) {
-                        mLastLoadedPeriods.add(0, periodIndex);
+                int index = mLastLoadedPeriods.indexOf(periodIndex);
+                if (index < 0 || forceIt) {
+                    mLastLoadedPeriods.add(0, periodIndex);
 
-                        int halfDaysInPeriod = mDaysInPeriod / 2;
+                    int halfDaysInPeriod = mDaysInPeriod / 2;
 
-                        Calendar now = Calendar.getInstance();
-                        now.add(Calendar.DATE, periodIndex * mDaysInPeriod - halfDaysInPeriod);
-                        now.set(Calendar.HOUR_OF_DAY, 0);
-                        now.set(Calendar.MINUTE, 0);
-                        now.set(Calendar.SECOND, 0);
-                        now.set(Calendar.MILLISECOND, 0);
+                    Calendar now = Calendar.getInstance();
+                    now.add(Calendar.DATE, periodIndex * mDaysInPeriod - halfDaysInPeriod);
+                    now.set(Calendar.HOUR_OF_DAY, 0);
+                    now.set(Calendar.MINUTE, 0);
+                    now.set(Calendar.SECOND, 0);
+                    now.set(Calendar.MILLISECOND, 0);
 
-                        Calendar then = Calendar.getInstance();
-                        then.add(Calendar.DATE, periodIndex * mDaysInPeriod + halfDaysInPeriod - 1);
-                        then.set(Calendar.HOUR_OF_DAY, 23);
-                        then.set(Calendar.MINUTE, 59);
-                        then.set(Calendar.SECOND, 59);
-                        then.set(Calendar.MILLISECOND, 999);
+                    Calendar then = Calendar.getInstance();
+                    then.add(Calendar.DATE, periodIndex * mDaysInPeriod + halfDaysInPeriod - 1);
+                    then.set(Calendar.HOUR_OF_DAY, 23);
+                    then.set(Calendar.MINUTE, 59);
+                    then.set(Calendar.SECOND, 59);
+                    then.set(Calendar.MILLISECOND, 999);
 
-                        Bundle args = new Bundle();
-                        args.putLong(ARG_CAL_LOAD_NOW, now.getTimeInMillis());
-                        args.putLong(ARG_CAL_LOAD_THEN, then.getTimeInMillis());
+                    Bundle args = new Bundle();
+                    args.putLong(ARG_CAL_LOAD_NOW, now.getTimeInMillis());
+                    args.putLong(ARG_CAL_LOAD_THEN, then.getTimeInMillis());
 
-                        //Log.d(TAG, String.format("loadPeriod %d(%d) %s - %s", periodIndex, index, SimpleDateFormat.getDateTimeInstance().format(now.getTime()), SimpleDateFormat.getDateTimeInstance().format(then.getTime())));
-                        if (index >= 0) {
-                            mLastLoadedPeriods.remove(index);
-                            getLoaderManager().restartLoader(periodIndex, args, CalendarFragment2.this);
-                        } else {
-                            getLoaderManager().initLoader(periodIndex, args, CalendarFragment2.this);
-                        }
+                    //Log.d(TAG, String.format("loadPeriod %d(%d) %s - %s", periodIndex, index, SimpleDateFormat.getDateTimeInstance().format(now.getTime()), SimpleDateFormat.getDateTimeInstance().format(then.getTime())));
+                    if (index >= 0) {
+                        mLastLoadedPeriods.remove(index);
+                        getLoaderManager().restartLoader(periodIndex, args, CalendarFragment2.this);
+                    } else {
+                        getLoaderManager().initLoader(periodIndex, args, CalendarFragment2.this);
+                    }
 
-                        while (mLastLoadedPeriods.size() > 3) {
-                            int removed = mLastLoadedPeriods.remove(mLastLoadedPeriods.size() - 1);
+                    while (mLastLoadedPeriods.size() > 3) {
+                        int removed = mLastLoadedPeriods.remove(mLastLoadedPeriods.size() - 1);
 
-                            //Log.d(TAG, String.format("period removed %d", removed));
+                        //Log.d(TAG, String.format("period removed %d", removed));
 
-                            getLoaderManager().destroyLoader(removed);
-                        }
+                        getLoaderManager().destroyLoader(removed);
                     }
                 }
-            }.run();
+            }).run();
         }
 
         @Override
