@@ -71,12 +71,7 @@ public class KusssContentProvider extends ContentProvider {
     private static final int CODE_CURRICULA = 7;
     private static final int CODE_CURRICULA_ID = 8;
 
-    private static final Comparator<String> TermComparator = new Comparator<String>() {
-        @Override
-        public int compare(String lhs, String rhs) {
-            return rhs.compareTo(lhs);
-        }
-    };
+    private static final Comparator<String> TermComparator = (lhs, rhs) -> rhs.compareTo(lhs);
 
     private static final UriMatcher sUriMatcher = new UriMatcher(
             UriMatcher.NO_MATCH);
