@@ -6,7 +6,7 @@
  *  \________|____|__ \______/   \____|__  /   __/|   __/
  *                   \/                  \/|__|   |__|
  *
- *  Copyright (c) 2014-2017 Paul "Marunjar" Pretsch
+ *  Copyright (c) 2014-2018 Paul "Marunjar" Pretsch
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -34,7 +34,6 @@ import java.util.Date;
 
 public class MensaDayTabItem extends SlidingTabItem {
 
-    private static final String DATE_PATTERN = "EEEE";
     private final Date mDate;
 
     public MensaDayTabItem(String title, Date date) {
