@@ -6,7 +6,7 @@
  *  \________|____|__ \______/   \____|__  /   __/|   __/
  *                   \/                  \/|__|   |__|
  *
- *  Copyright (c) 2014-2017 Paul "Marunjar" Pretsch
+ *  Copyright (c) 2014-2018 Paul "Marunjar" Pretsch
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -27,6 +27,7 @@ package org.voidsink.anewjkuapp.fragment;
 
 import org.voidsink.anewjkuapp.mensa.MenuLoader;
 import org.voidsink.anewjkuapp.mensa.RaabMenuLoader;
+import org.voidsink.anewjkuapp.utils.Consts;
 
 public class MensaRaabFragment extends MensaFragmentDetail {
 
@@ -35,4 +36,8 @@ public class MensaRaabFragment extends MensaFragmentDetail {
         return new RaabMenuLoader();
     }
 
+    @Override
+    protected String getMensaLink() {
+        return Consts.MENSA_MENU_RAAB;
+    }
 }
