@@ -6,7 +6,7 @@
  *  \________|____|__ \______/   \____|__  /   __/|   __/
  *                   \/                  \/|__|   |__|
  *
- *  Copyright (c) 2014-2017 Paul "Marunjar" Pretsch
+ *  Copyright (c) 2014-2018 Paul "Marunjar" Pretsch
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -38,8 +38,6 @@ import java.util.Date;
 import java.util.Locale;
 
 public class Curriculum {
-
-    private static final String TAG = Curriculum.class.getSimpleName();
 
     private boolean mIsStandard;
     private String mCid;
