@@ -76,7 +76,7 @@ public class Consts {
     public static final String SCREEN_LOGIN = "/login";
     public static final String SCREEN_DASHCLOCK = "/dashclock";
 
-    // Laoder IDs
+    // Loader IDs
     public static final int LOADER_ID_COURSES = 1;
     public static final int LOADER_ID_ASSESSMENTS = 2;
 
@@ -85,6 +85,12 @@ public class Consts {
     public static final String CHANNEL_ID_GRADES = BuildConfig.APPLICATION_ID + ".GRADES";
     public static final String CHANNEL_ID_EXAMS = BuildConfig.APPLICATION_ID + ".EXAMS";
     public static final String CHANNEL_ID_DEFAULT = BuildConfig.APPLICATION_ID + ".DEFAULT";
+
+    // URLs
+    public static final String MENSA_MENU_CLASSIC = "http://menu.mensen.at/index/index/locid/1";
+    public static final String MENSA_MENU_CHOICE = "http://menu.mensen.at/index/index/locid/1";
+    public static final String MENSA_MENU_KHG = "https://www.dioezese-linz.at/institution/8075/essen/menueplan";
+    public static final String MENSA_MENU_RAAB = "http://www.sommerhaus-hotel.at/de/linz#speiseplan";
 
     private Consts() {
     }
